Fixing linting changes
