+++ pipelines/zendesk/helpers/credentials.py
@@ -4,7 +4,6 @@ This module handles how credentials are read in dlt sources
 from dlt.common.configuration.specs import CredentialsConfiguration
 from dlt.common.configuration import configspec
 from dlt.common.typing import TSecretValue
-from typing import Optional
 
 
 @configspec
